+++ homework/src/HW1/submissions/partA/vincentsaw/facebook/Post.java
@@ -1,0 +1,5 @@
+package HW1.submissions.partA.vincentsaw.facebook;
+
+public interface Post{
+    void display();
+}
