--- homework/src/HW1/submissions/partA/vincentsaw/facebook/Post1.java
@@ -1,14 +1,0 @@
-package HW1.submissions.partA.vincentsaw.facebook;
-
-//This was the original post class, that was implemented for our "Facebook".
-
-class Post1 {
-    private String text;
-
-    // Constructor, getters, and setters for text
-
-    void display() {
-        System.out.println(text);
-        // Code for displaying text post
-    }
-}
